put bubble styles into a separate module
