+++ src/components/Game.tsx
@@ -1,40 +1,7 @@
 import React, { useContext } from 'react'
 import { GameContext } from '../contexts/GameContext'
 import { Bubble, GameProps } from '../types/GameTypes'
-
-const bubbleStyle = (bubble: Bubble): Object => {
-  return ({
-    backgroundColor: bubble.color,
-    background: `radial-gradient(circle at 20px 20px, ${bubble.color}, rgba(0,0,0,0.7))`
-  })
-}
-
-const bubbleAnimation = (shoot: boolean, hitCoordinates: number[]): Object => {
-  const hexagonalCorrection = hitCoordinates[0] % 2 * 25
-  const x = hitCoordinates[1] * 50 + hexagonalCorrection
-  const y = 500 - hitCoordinates[0]*50
-
-  if(shoot){
-    return ({
-      position: 'absolute',
-      bottom: y,
-      left: x,
-      transition: 'all 0.5s' 
-    })
-  }
-  return ({
-    position: 'absolute',
-    bottom: '0px',
-    left: '250px',
-    transition: 'all 0s' 
-  })
-}
-
-const shootingBubbleStyle = (shoot: boolean, bubble: Bubble, hitCoordinates: number[]): Object => {
-  const color = bubbleStyle(bubble)
-  const position = bubbleAnimation(shoot, hitCoordinates)
-  return {...color, ...position} 
-}
+import { bubbleStyle, shootingBubbleStyle } from './BubbleStyles'
 
 const Game: React.FC<GameProps> = ({handleMousePosition, handleMouseDown, handleTransitionEnd, angle, shoot}) => {
   const { state }: any = useContext(GameContext)
